Added new copy in about and allowed getting back to home page
Fixed issue #53 and issue #50.  Updated the about page with the new copy provided by @meatherly and linked back to the "Home" page with the information paragraph using the "Nashville Transparency Project" in the upper lefthand corner. It is dynamically regenerating content with Javascript - I feel like we could do better if anyone has time.