Removed income brackets from 'Census Predicted' charts
Removed the repetitious income brackets from the 'Census Predicted' charts (e.g. 'Lower Income (Less than $33,000).  This makes sense in the desktop version, but on smaller widths the columns get smashed into a single column, so we might want to do some bootstrap magic to add the title, but hide it until we get down to a single column.