Min-width for graphs and cleared selectors at top
Updated the site to improve mobile responsiveness. Tested it using Chrome's mobile emulators, but proof is in the pudding, so there may be more fixes down the line once we deploy and test the real deal.

Wrapped line-graphs in a row class div to give it a min width.  Added Bootstrap col classes to the selector container to clear it from the "About" message.

Cleaned up the markup, removing extraneous divs and css, and made some of the ids more meaningful. For example, replaced `id='charts-container2'` with `id='census-pie-charts`. Organized css selectors in order of descending specificity, and alphabetically.