Convert to static site
This converst the application to a static site, moving all the data processing and munging to a single script, import_data.py.

All references to Flask, RethinkDB and CherryPy are removed.  All js is now vendored into the build to prepare for bundling the app.

It works, but it's still not fully automated. See TODOs for examples.

TODOS:
* Build line charts in import data. Right now these are hardcoded from who knows where in `graphs
* Test build against a new dataset from Socrata
* Use an honest static site generator so we aren't manually copying json into javascript files (we can template in.
* Don't regenerate data that doesn't change in "Summary of Demographics Over Time". See `aboutPage()`