Use Docker instead of Vagrant
The Vagrantfile is based on an old build from @neolytics, which in turn is based on a private box from NextGxDx.  Use Dockerfile and docker-compose so that we can simply run `docker-compose up`.

This also involved updating our rethinkdb connections to read from environment variables populated by the docker-compose process, and update one query to work with the latest rethinkdb version.

I've created a `codefornashville` organization Docker Hub and added the output of this Dockerfile.  I'd like to add a few more members of Code for Nashville so that its not just me who can update the Dockerfile. Right now to update the build we need to push up a new Docker image each time we change our requirements.

I removede dependencies.sh because it just install an out of date Vagrant, and replaced Vagrant with Docker instructions.

I'm not able to get autocomplete for "Departments" in the "Custom Reports" section.