Replace import_data.py with JS equivalent
This replaces the Python script that normalizes and transforms the input data with a JS equivalent.  Because it was easier than re-implementing the complex backend code, this also simplifies the data structure we need to generate, and makes appropriate compensations in the frontend, where required.

Because this has a bit of branching logic, I added commander, a simple and helpful library to create command line tools.  This matches the use of argparse in the Python version.

The previous version of the script pre-generated all possible combinations of department/gender|ethnicity/income-level totals. To simplify, the JS version just normalizes the input of the latest date into another flat CSV, so all 9000 of the latest employees are loaded by the client, instead of summaries.  This is about a 3X increase in data-size (summary-by-department.json 30KB, employees.csv 82KB). This makes the client much more flexible - we can easily modify or create new visualizations w/out pushing it into the import script. To support loading CSVs, this also adds d3-dsv, a handy and reliable library for parsing delimited files from the popular d3 ecosystem.

A similar change was made to the summaries, which have ethnicity totals for each year, for each income level.  The backend code is simplified to generate a total per department/ethnicity/date. These are sent as flat data, which lets the client decide how to index and organize, rather than assuming the best structure for visualization.  On examination, HighCharts exposed a simpler API for submitting line chart data, which this flat format was ablet to adapt to easily. The format change was from expecting {date: [], items: []} pairs to mapping each item (which has a date attribute) into a [date, value] pair. See https://www.highcharts.com/docs/chart-concepts/series for more info on submitting data to a series.

This slightly changes naming and makes it consistent - "IncomeCategory" is renamed "SalaryBucketId", and ethnic/demographic has been corrected to consistently be "ethnicity".

This removes the last bit of Python from the project.  This lowers the barrier to entry and maintainance burden, and unblocks deploying the update script as an AWS lambda function.