Made old input files comma separated instead of "|" separated
Used the following little script to convert our old "|" separated files in to "," separated files.

```python
def swap_delimiter(filepath, old_delimiter, new_delimiter=','):
     with open(filepath, 'r') as f:
         reader = csv.DictReader(f, delimiter=old_delimiter)
         rows = list(reader)

     with open(filepath, 'w') as f:
         writer = csv.DictWriter(f, rows[0].keys(), delimiter=new_delimiter)
         writer.writeheader()
         writer.writerows(rows)
```

The old python code "sniffed" out the delimeter by reading the first 2KB of the file. This removes the need for it to do this - and the future node implementation which doesn't seem to come with this feature.