Fetch data instead of importing it
Close #132. Fetch data (static json files, not an API) instead of importing it with Webpack and compiling it into the static build.  While the latter is more efficient, fetching it gives us flexibility to store files in a separate server, which we need to write and read files from S3.

This changes the import_data script to write to our `public/data` folders where they can be fetched by the client.  Components have been modified to fetch the data in the public folder instead of importing it from source. This sets minimal error handling (console.error) which could and should be improved.