Issue 129: Generate all dates on import
This modifies the import to generate modified CSVs for all dates on import.  This now saves each dates employees within a folder of the form YYYY-MM-DD.  This supports a UI that can switch between different years of the report.

Instead of parsing and modifying the date, this converts it once to a relatively parseable format (YYYY-MM-DD) an uses that for all subsequent usages.

To minimize calls to `setState`, the pattern of fetchind date has been changed from directly calling `setState` in the result of the fetch promise to returning the values, and having `componentDidMount` call `setState` with the results of any fetched data.