Add date selector to ExploreCharts
This adds a date selector, taking advantage of the initial work we did exporting the employees for each date.

This also adds labels and reformats so that all three input selectors are easier to view. It also utilizes the `clearable` and `searchable` attributes to remove unnecessary options (`clearable` and `searchable` for the two element **Attribute** dropdown.