Add date select to "Explore" tab
Closes #129. This adds a date selector to tab Explore, taking advantage of some initial refactoring where we export the employees for all dates instead of jus the latest.

This modifies the import to generate modified CSVs for all dates on import. This now saves each dates employees within a folder of the form YYYY-MM-DD. As a touch-up, instead of parsing and modifying the date, this converts it once to a relatively parseable format (YYYY-MM-DD) an uses that for all subsequent usages. Processing all the dates every time we call the script takes a second or two, so I removed the option for a particular date, and removed commander, which we used to declaratively provide that option.

Calls to setState that result from fetches have been grouped using Promise.all to wait for all to complete.

This also adds labels and gives each select on explore that all three input selectors are easier to view and understand. Since the clearable attribute was required for the new date select, I also added it to the Attribute select. I noticed the searchable attribute to prevent users from auto-
completing, and also added it to the two-element Atribute dropdown.