Change data source depending on environment (prod, dev)
Use built-in NODE_ENV to change the data source for local development.

If the app is started using `npm build`, the environment will be
considered production, and the S3 buckets should be used as the data
sources.

If it is started using `npm start` or `npm test`, it will be considered
a non-production environment, and will use the local data in
/public/test/data rather than trying to pull chart data from S3.

Required to allow people to run the site locally without CORS getting in
the way.