Add test data for non-production testing
Due to CORS restrictions, we can't run S3 testing on local machines, so
we need to fudge the data sources using local files.

As the production site continues to pull data, these test files will not
represent the complete set of production data, but that's not really a
problem for functionality testing (unless the S3 component is what's
being tested, in which case we've got a whole other set of issues).